requirements_dev.txt
python-dotenv
transformers==4.34.1
accelerate==0.21.0
bitsandbytes==0.41.1
sentence_transformers==2.2.2
spacy==3.7.2
openai
llama-index==0.8.55
llama_hub==0.0.43
langchain==0.0.329
pypdf
"unstructured[pdf]"
pymupdf==1.23.5
llmsherpa==0.1.3
ipywidgets==7.7.1
gradio
